:shower: move shared method to library
