+++ app/lib/messages.js
@@ -3,7 +3,7 @@ function invalidPostcodeMessage(location) {
 }
 
 function emptyPostcodeMessage() {
-  return 'You must insert a postcode to find a pharmacy.';
+  return 'You must insert a place or a postcode to find a pharmacy.';
 }
 
 function technicalProblems() {
